pyproject.toml
[tool.poetry]
name = "authtelegrambot"
version = "0.1.0"
description = ""
authors = ["BiqRedMac <redshine.nik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
